Moves interface, enum etc. into a new package.
