+++ src/main/java/com/alexfu/sqlitequerybuilder/api/Builder.java
@@ -1,4 +1,4 @@
-package com.alexfu.sqlitequerybuilder;
+package com.alexfu.sqlitequerybuilder.api;
 
 public interface Builder {
   public String build();
