+++ src/main/java/com/alexfu/sqlitequerybuilder/api/CreateTable.java
@@ -1,4 +1,4 @@
-package com.alexfu.sqlitequerybuilder;
+package com.alexfu.sqlitequerybuilder.api;
 
 /**
  * @author Steven Wu
